Cleanup, time to write tests and documentation.
