--- src/Common/Winterhaven.Common.DTOs/Players/PlayerActionDto.cs
@@ -1,17 +1,0 @@
-﻿namespace Winterhaven.Common.DTOs.Players;
-
-/// <summary>
-/// </summary>
-/// <param name="Type">
-/// </param>
-/// <param name="MoveX">
-/// </param>
-/// <param name="MoveY">
-/// </param>
-/// <param name="Identifier">
-/// </param>
-public sealed record PlayerActionDto(
-    string Type,
-    double MoveX,
-    double MoveY,
-    double Identifier);
