--- src/Common/Winterhaven.Common.DTOs/Winterhaven.Common.DTOs.csproj
@@ -1,1 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" />
