--- src/Common/Winterhaven.Common.Extensions/Winterhaven.Common.Extensions.csproj
@@ -1,6 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <ItemGroup>
-	<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" />
-	<PackageReference Include="Microsoft.Extensions.Options.DataAnnotations" />
-  </ItemGroup>
-</Project>
