+++ src/Common/Winterhaven.Common/DTOs/Players/PlayerActionDto.cs
@@ -1,0 +1,12 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Winterhaven.Common.DTOs.Players;
+
+/// <summary>
+/// </summary>
+[ExcludeFromCodeCoverage]
+public sealed record PlayerActionDto(
+    string Type,
+    double MoveX,
+    double MoveY,
+    double Identifier);
