+++ src/Common/Winterhaven.Common/Events/IEvent.cs
