+++ src/Common/Winterhaven.Common/Events/PublishOptions.cs
