+++ src/Common/Winterhaven.Common/Events/SubscribeOptions.cs
