+++ src/Common/Winterhaven.Common/Events/Users/UserLoggedOutEvent.cs
