+++ src/Common/Winterhaven.Common/Exceptions/AuthorizationException.cs
