+++ src/Common/Winterhaven.Common/Exceptions/MessageBusException.cs
@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Winterhaven.Brokering.Exceptions;
 
 /// <summary>
 /// </summary>
+[ExcludeFromCodeCoverage]
 public sealed class MessageBusException : Exception
 {
     /// <summary>
@@ -14,8 +16,6 @@ public sealed class MessageBusException : Exception
 
     /// <summary>
     /// </summary>
-    /// <param name="message">
-    /// </param>
     public MessageBusException(string message)
         : base(message)
     {
@@ -23,10 +23,6 @@ public sealed class MessageBusException : Exception
 
     /// <summary>
     /// </summary>
-    /// <param name="message">
-    /// </param>
-    /// <param name="innerException">
-    /// </param>
     public MessageBusException(string message, Exception innerException)
         : base(message, innerException)
     {
