+++ src/Common/Winterhaven.Common/Exceptions/ValidationException.cs
