+++ src/Directory.Build.props
@@ -9,8 +9,6 @@
 		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
 		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-		<BaseOutputPath>..\bin\$(MSBuildProjectName)</BaseOutputPath>
-		<BaseIntermediateOutputPath>..\obj\$(MSBuildProjectName)</BaseIntermediateOutputPath>
 	</PropertyGroup>
 
 	<PropertyGroup>
