+++ src/Events/Winterhaven.Events.Tests/Winterhaven.Events.Tests.csproj
@@ -20,6 +20,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Winterhaven.Brokering.NATS\Winterhaven.Brokering.NATS.csproj" />
+    <ProjectReference Include="..\Winterhaven.Events\Winterhaven.Events.csproj" />
   </ItemGroup>
 </Project>
