+++ src/Events/Winterhaven.Events/Winterhaven.Events.csproj
@@ -1,20 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
     <PackageReference Include="NATS.Net" />
   </ItemGroup>
-
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\Winterhaven.Common\Winterhaven.Common.csproj" />
+  </ItemGroup>
   <ItemGroup>
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>Winterhaven.Brokering.NATS.Tests</_Parameter1>
+      <_Parameter1>Winterhaven.Events.Tests</_Parameter1>
     </AssemblyAttribute>
+    <!-- For strong-named Castle (NSubstitute ships both; keep both entries) -->
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
       <_Parameter1>DynamicProxyGenAssembly2, PublicKey=0024000004800000940000000602000000240000525341310004000001000100c547cac37abd99c8db225ef2f6c8a3602f3b3606cc9891605d02baa56104f4cfc0734aa39b93bf7852f7d9266654753cc297e7d2edfe0bac1cdcf9f717241550e0a7b191195b7667bb4f64bcb8e2121380fd1d9d46ad2d92d2d15605093924cceaf74c4861eff62abf69b9291ed0a340e113be11e6a7d3113e92484cf7045cc7</_Parameter1>
     </AssemblyAttribute>
-  </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Winterhaven.Brokering\Winterhaven.Brokering.csproj" />
+    <!-- For unsigned Castle (needed when the host assembly is NOT strong-named) -->
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
+      <_Parameter1>DynamicProxyGenAssembly2</_Parameter1>
+    </AssemblyAttribute>
   </ItemGroup>
 
 </Project>
+
