+++ src/Gateway/Winterhaven.Gateway.Integrations/Winterhaven.Gateway.Integrations.csproj
@@ -11,7 +11,6 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="NUnit" />
     <PackageReference Include="NUnit.Analyzers">
@@ -23,7 +22,6 @@
     </PackageReference>
     <PackageReference Include="Testcontainers.Nats" />
     <PackageReference Include="WireMock.Net" />
-    <PackageReference Include="Microsoft.IdentityModel.Tokens" />
     <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
   </ItemGroup>
