+++ src/Gateway/Winterhaven.Gateway.Presentation/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Winterhaven.Brokering.NATS.Extensions;
+using Winterhaven.Events.Extensions;
 using Winterhaven.Gateway.Infrastructure.Extensions;
 using Winterhaven.Gateway.Presentation.Extensions;
 using Winterhaven.Gateway.Presentation.Middleware;
@@ -82,7 +82,7 @@ internal sealed class Startup
 
         services.AddEventForwarder<PlayerEventForwarder>();
 
-        services.AddBrokeringServices(Configuration);
+        services.AddEventServices(Configuration);
         services.AddGatewayInfrastructureServices(Configuration);
 
         services.AddHttpContextAccessor();
